Unfied handling of the migration table
